+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 - added angle param to add/cut_rect()
 - added add/cut_slot variant with 2 points
 - added add/cut/intersect_sketch()
-- added make_sphere()
+- added make_sphere() and make_cylinder()
 
 ## [0.4] - 2024-01-12